Move screens to individual screen packages
